requirements.txt
fastapi
uvicorn
openai
python-dotenv
langchain
langchain-openai
langchain_community